NewsAgent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.0.3
numpy==1.24.3
google-cloud-storage==2.10.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pyarrow==12.0.1
protobuf==3.20.3
google-generativeai==0.3.2
newspaper3k==0.2.8
beautifulsoup4==4.12.2
nltk==3.8.1
tf-keras==2.15.0
transformers==4.35.2
torch==2.1.0
huggingface_hub==0.25.2
